% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svp_internal.R
\name{correction_factor_ged_approx}
\alias{correction_factor_ged_approx}
\title{Approximate correction factor C_g(nu) for GED leverage (diagnostic use only)}
\usage{
correction_factor_ged_approx(nu, n_nodes = 200L)
}
\arguments{
\item{nu}{GED shape parameter (nu > 0)}

\item{n_nodes}{Number of GH quadrature nodes (default 200)}
}
\value{
Approximate C_g(nu)
}
\description{
\eqn{C_g(\nu) = E[|u|] \cdot \textrm{Cov}(z, F_{GED}^{-1}(\Phi(z))) / \sqrt{2/\pi}}.
This is a first-order approximation; estimation uses the exact implicit equation.
}
\keyword{internal}
