% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svp_internal.R
\name{correction_factor_t}
\alias{correction_factor_t}
\title{Correction factor C_t(nu) for Student-t leverage estimation
Under scale mixture u_t = z_t * lambda_t^{-1/2}, C_t(nu) = [E[lambda^{-1/2}]]^2
= (nu/2) * [Gamma((nu-1)/2) / Gamma(nu/2)]^2.  Exact, parameter-free.}
\usage{
correction_factor_t(nu)
}
\arguments{
\item{nu}{Degrees of freedom (nu > 1)}
}
\value{
C_t(nu), always > 1 for finite nu (approaches 1 as nu -> Inf)
}
\description{
Correction factor C_t(nu) for Student-t leverage estimation
Under scale mixture u_t = z_t * lambda_t^{-1/2}, C_t(nu) = [E[lambda^{-1/2}]]^2
= (nu/2) * [Gamma((nu-1)/2) / Gamma(nu/2)]^2.  Exact, parameter-free.
}
\keyword{internal}
