% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman.R
\name{density_eps_ged}
\alias{density_eps_ged}
\title{Density of Centered log-GED^2 Measurement Noise}
\usage{
density_eps_ged(y, nu)
}
\arguments{
\item{y}{Numeric vector. Evaluation points (centered: E[eps] = 0).}

\item{nu}{Numeric. GED shape parameter.}
}
\value{
Density values.
}
\description{
Density of Centered log-GED^2 Measurement Noise
}
\keyword{internal}
