% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svp_internal.R
\name{.compute_EH}
\alias{.compute_EH}
\title{Compute EH cross-moment for leverage estimation
EH = demeaned sample cross-moment (1/(T-2)) * sum((|y_t| - mean|y|)(y_{t-1} - mean(y)))}
\usage{
.compute_EH(y, rho_type = "pearson")
}
\arguments{
\item{y}{Numeric vector of observations}

\item{rho_type}{"pearson" or "kendall"}
}
\value{
EH value
}
\description{
Compute EH cross-moment for leverage estimation
EH = demeaned sample cross-moment (1/(T-2)) * sum((|y_t| - mean|y|)(y_{t-1} - mean(y)))
}
\keyword{internal}
