% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svp_internal.R
\name{.get_gh}
\alias{.get_gh}
\title{Get cached Gauss-Hermite nodes/weights for N(0,1)}
\usage{
.get_gh(n = 200L)
}
\arguments{
\item{n}{Number of nodes (default 200)}
}
\value{
List with nodes and weights
}
\description{
Get cached Gauss-Hermite nodes/weights for N(0,1)
}
\keyword{internal}
