% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svp_internal.R
\name{ged_E_abs_u}
\alias{ged_E_abs_u}
\title{E[|u|] for standardized GED(nu) with Var = 1
Closed form: sqrt(Gamma(1/nu)/Gamma(3/nu)) * Gamma(2/nu) / Gamma(1/nu)}
\usage{
ged_E_abs_u(nu)
}
\arguments{
\item{nu}{GED shape parameter (nu > 0)}
}
\value{
E[|u|]
}
\description{
E[|u|] for standardized GED(nu) with Var = 1
Closed form: sqrt(Gamma(1/nu)/Gamma(3/nu)) * Gamma(2/nu) / Gamma(1/nu)
}
\keyword{internal}
