% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{laggedMat}
\alias{laggedMat}
\title{Create Lagged Matrix from a Time Series}
\usage{
laggedMat(y, lags, balance = TRUE)
}
\arguments{
\item{y}{Numeric vector. The time series.}

\item{lags}{Integer. Number of lags to include.}

\item{balance}{Logical. If \code{TRUE}, remove rows with NAs.}
}
\value{
A matrix with columns \code{y(t), y(t-1), ..., y(t-lags)}.
}
\description{
Create Lagged Matrix from a Time Series
}
\keyword{internal}
