% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svp_internal.R
\name{LRT_moment_lev_ged}
\alias{LRT_moment_lev_ged}
\title{GMM moments for SVL(p)-GED with fixed A matrix (p+4 conditions)
Uses exact GED leverage moment.}
\usage{
LRT_moment_lev_ged(y, mdl_out, Amat, rho_type, del = 1e-10)
}
\description{
GMM moments for SVL(p)-GED with fixed A matrix (p+4 conditions)
Uses exact GED leverage moment.
}
\keyword{internal}
