% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svp_internal.R
\name{LRT_moment_lev_t}
\alias{LRT_moment_lev_t}
\title{GMM moments for SVL(p)-Student-t with fixed A matrix (p+4 conditions)}
\usage{
LRT_moment_lev_t(y, mdl_out, Amat, rho_type, del = 1e-10)
}
\description{
GMM moments for SVL(p)-Student-t with fixed A matrix (p+4 conditions)
}
\keyword{internal}
