% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svp_internal.R
\name{LRT_moment_lev_t_Amat}
\alias{LRT_moment_lev_t_Amat}
\title{GMM moments for SVL(p)-Student-t with HAC weighting (p+4 conditions)}
\usage{
LRT_moment_lev_t_Amat(y, mdl_out, rho_type, del = 1e-10, Bartlett = TRUE)
}
\description{
GMM moments for SVL(p)-Student-t with HAC weighting (p+4 conditions)
}
\keyword{internal}
