% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svp_internal.R
\name{qged_std}
\alias{qged_std}
\title{Quantile function for standardized GED(nu) with Var = 1
Uses the relationship between GED CDF and incomplete gamma function.}
\usage{
qged_std(p, nu)
}
\arguments{
\item{p}{Probability (0 < p < 1). Clamped to [1e-15, 1-1e-15] internally.}

\item{nu}{GED shape parameter (nu > 0)}
}
\value{
Quantile value
}
\description{
Quantile function for standardized GED(nu) with Var = 1
Uses the relationship between GED CDF and incomplete gamma function.
}
\keyword{internal}
